backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic-settings
bcrypt
python-jose[cryptography]
pytest
httpx
email-validator
python-magic
python-multipart
pypdf2
ollama
